remove navigations and refactor print templates
