--- codbex-orders/navigations/purchase-orders/service.js
@@ -1,17 +1,0 @@
-const navigationData = {
-    id: 'purchase-order-navigation',
-    label: "Purchase Orders",
-    group: "purchasing",
-    order: 100,
-    link: "/services/web/codbex-orders/gen/codbex-orders/ui/PurchaseOrder/index.html?embedded"
-};
-
-function getNavigation() {
-    return navigationData;
-}
-
-if (typeof exports !== 'undefined') {
-    exports.getNavigation = getNavigation;
-}
-
-export { getNavigation }
