--- codbex-orders/navigations/sales-order-total-report/navigation.extension
@@ -1,6 +1,0 @@
-{
-    "extensionPoint": "dashboard-navigations",
-    "module": "codbex-orders/navigations/sales-order-total-report/service.js",
-    "description": "Sales Order Total Report Navigation",
-    "role": "codbex-orders.entities.SalesOrdersTotalReportFilterReadOnly,codbex-orders.Reports.SalesOrdersTotalReportReadOnly"
-}
