--- codbex-orders/navigations/sales-orders/navigation.extension
@@ -1,6 +1,0 @@
-{
-    "extensionPoint": "dashboard-navigations",
-    "module": "codbex-orders/navigations/sales-orders/service.js",
-    "description": "Sales Order Navigation",
-    "role": "codbex-orders.SalesOrder.SalesOrderReadOnly,codbex-orders.SalesOrder.SalesOrderFullAccess"
-}
