+++ codbex-orders/print/SalesOrder/view.js
@@ -1,7 +1,7 @@
 const viewData = {
     id: 'sales-order-print',
     label: 'Print',
-    link: '/services/ts/codbex-templates/print/sales-order-print-template.ts',
+    path: '/services/ts/codbex-templates/print/sales-order-print-template.ts',
     perspective: 'SalesOrder',
     view: 'SalesOrder',
     type: 'entity',
@@ -9,7 +9,5 @@ const viewData = {
 };
 
 if (typeof exports !== 'undefined') {
-    exports.getDialogWindow = function () {
-        return viewData;
-    }
+    exports.getView = () => viewData;
 }
